// src/app/api/model/plant.ts
/**
 * P.L.A.N.T.S. API
 * Provides functions to interact with P.L.A.N.T.S. systems
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Plant { 
    microbit: number;
    network?: string;
    description?: string;
    connected: boolean;
    sink: boolean;
    sensors: Array<string>;
}